requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy>=1.4.0
asyncpg>=0.30.0
alembic>=1.7.3
pydantic_settings>=2.7.1